Move install under pkg/
